requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
matplotlib>=3.7
pyyaml>=6.0
tqdm>=4.65
